Add new param to allow/disallow JSON-RPC Notification (#385)
A Notification is a Request object without an "id" member. A Request
object that is a Notification signifies the Client's lack of interest
in the corresponding Response object, and as such, no Response object
needs to be returned to the client. The Server MUST NOT reply to a
Notification, including those that are within a batch request.

The new parameter allows defining whether the method will support the
Notification Response object, the default is true. The use of this
parameter is:

```
@jsonrpc.method('App.method', notification: bool = True)
```

When `notification` parameter is false, the method itself doesn't allow
the Notification Response object, in the other word, if the client calls
the method with a Notification Request object the server will respond
with an Error object with code = -32600 and message = Invalid Request.

Resolves: #369