Add tests to the Python build, wheel, and Mypyc packages.
Create shared tests to be used for both unit and integration
testing, and increase the coverage of unit tests to 100%. Additionally,
before publishing the pre-release, run the shared tests against the
build version for each OS supported by the code.

The intention here is to ensure that all build packages are tested
and function correctly for all supported features.

Resolves: #577