Enhance the error handler mechanism
Add the capability to change the HTTP status code and HTTP headers of response
when an error is handled. Also, if no specific error handler found, use the
generic ServerError handler if available, it allows the frameworks and developers
to debug and catch unexpected errors with context.

When the generic ServerError handler is triggered, the original exception that
caused this unexpected error is injected at the attribute `original_exception` and
it will be available on error handler of ServerError.